RDKB-57868: Update configuration change appropriately in non-db case
Reason for change: In case of OneWifi interfacing with easymesh agent,
the changed configuration was not updating appropriately. This
resulted in radio config encoding stale data and notifying older
information. This commit addresses this issue.

Test Procedure: Ensured that the config change is reflected in radio
subdoc as part of the channel change request to agent.

Risks: Low
Priority: P1
Signed-off-by: Amarnath Hullur Subramanyam <amarnath_hullursubramanyam@comcast.com>
Change-Id: Iaab6a52692e1c4f6c3d002932e2cc030f36344fd