RDKB-58038: Give precedence to channel in RadioConfig.
Reason for change: When there is a mismatch in the channel value
of subdoc in Radioconfig and currentoperatingclass, give precedence
to channel in Radioconfig to ensure backward compatibility.

Test Procedure: Verified the fix working with channel change only in
Radioconfig and ensured passing of the earlier failing testcase.

Risks: Low
Priority: P1
Signed-off-by: Amarnath Hullur Subramanyam <amarnath_hullursubramanyam@comcast.com>
Change-Id: I5e5f8fe671a3e6c67c7da827d1b1eb0e60ca70b0