RDKB-57741, RDKB-57868, RDKB-58478, RDKB-58559 : Send Assoc frame data in associated clients subdoc (#322)
* RDKB-57741: Send Assoc frame data in associated clients subdoc

Association/Reassociation frame data is encoded into a hex string
and sent as part of Associated Clients subdoc for the easy mesh
agent consumption.

Reason for change: Easymesh agent to be aware of associated client
data.
Test Procedure: Ensured that appropriate hex string is being sent
when assoc frame is received.
Risks: Low
Priority: P0

Change-Id: Ibe3ee428649efa6d925b9349e98052e28afd5577
Signed-off-by: Amarnath Hullur Subramanyam <amarnath_hullursubramanyam@comcast.com>

* RDKB-57868: Update configuration change appropriately in non-db case

Reason for change: In case of OneWifi interfacing with easymesh agent,
the changed configuration was not updating appropriately. This
resulted in radio config encoding stale data and notifying older
information. This commit addresses this issue.

Test Procedure: Ensured that the config change is reflected in radio
subdoc as part of the channel change request to agent.

Risks: Low
Priority: P1
Signed-off-by: Amarnath Hullur Subramanyam <amarnath_hullursubramanyam@comcast.com>
Change-Id: Iaab6a52692e1c4f6c3d002932e2cc030f36344fd

* RDKB-58478: Frame Data being absent in client subdoc not to be treated as error

Reason for change: Frame Data absent in client subdoc is a valid scenario
and should not be treated as error in decode. Frame Data will not be present
in case of disassoc event.
Test Procedure: Ensured that the Error messages are not printed if framedata
is absent.

* RDKB-58559: Address crash seen in stringtohex function

Reason for change: Crash was due to buffer overflow in accessing
the input buffer.
Test Procedure: Checked connection and assoc data being encoded
and decoded appropriately with easymesh agent and controller
Risks: Low
Priority: P1

---------

Signed-off-by: Amarnath Hullur Subramanyam <amarnath_hullursubramanyam@comcast.com>
Co-authored-by: Amarnath Hullur Subramanyam <amarnath.hs@gmail.com>
Co-authored-by: Amarnath Hullur Subramanyam <182549733+amarnathhullur@users.noreply.github.com>